@@ -2,7 +2,7 @@
     "name": "hexlet/code",
     "description": "PHP project on Laravel",
     "require": {
-        "php": "^7.3|^8.0",
+        "php": "7.4",
         "doctrine/dbal": "^3.1",
         "fideloper/proxy": "^4.4",
         "fruitcake/laravel-cors": "^2.0",
